+++ xcube_smos/nckcindex/constants.py
@@ -1,9 +1,23 @@
-DEFAULT_INDEX_NAME = "nckc-index"
-DEFAULT_SOURCE_PROTOCOL = "file"
+# The MIT License (MIT)
+# Copyright (c) 2023 by the xcube development team and contributors
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
 
-S3_PROTOCOL = "s3"
-EXAMPLE_S3_ENDPOINT_URL = "https://s3.cloudferro.com"
-EXAMPLE_S3_BUCKET = "EODATA"
-
-INDEX_CONFIG_VERSION = 2
 INDEX_CONFIG_FILENAME = "nckc-config.json"
+INDEX_CONFIG_VERSION = 2